+++ SharpSqlBuilder/Maps/SqlFilterItem.cs
@@ -1,7 +1,12 @@
-﻿namespace SharpSqlBuilder.Maps
+﻿using SharpSqlBuilder.Operands;
+using SharpSqlBuilder.Operators;
+
+namespace SharpSqlBuilder.Maps
 {
-    public class SqlFilterItem
+    public class SqlFilterItem: IOperable
     {
         public string PropertyName { get; set; }
+        public Operand AsOperand => new SqlFilterOperand(this);
     }
+
 }