+++ SharpSqlBuilder/Operators/IOperable.cs
@@ -1,0 +1,9 @@
+﻿using SharpSqlBuilder.Operands;
+
+namespace SharpSqlBuilder.Operators
+{
+    public interface IOperable
+    {
+        Operand AsOperand { get; }
+    }
+}